requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pandas==2.2.2
scikit-learn==1.5.2
matplotlib==3.9.2
sqlmodel==0.0.21
pydantic==2.9.0
python-dotenv==1.0.1
pytest==8.3.2
ruff==0.6.8
black==24.8.0
joblib==1.4.2
